+++ src/mob_proc.cpp
@@ -12,7 +12,7 @@
  *  This is free software and you are benefitting.  We hope that you       *
  *  share your changes too.  What goes around, comes around.               *
  ***************************************************************************/
-/* $Id: mob_proc.cpp,v 1.2 2002/06/13 04:41:08 dcastle Exp $ */
+/* $Id: mob_proc.cpp,v 1.3 2002/06/20 21:39:36 pirahna Exp $ */
 #ifdef LEAK_CHECK
 #include <dmalloc.h>
 #endif
@@ -3347,7 +3347,7 @@ int pet_shops(struct char_data *ch, int cmd, char *arg)
     if (cmd==59) { /* List */
 	send_to_char("Available pets are:\n\r", ch);
 	for(pet = world[pet_room].people; pet; pet = pet->next_in_room) {
-	    sprintf(buf, "%8ld - %s\n\r",
+	    sprintf(buf, "%8lld - %s\n\r",
 		3*GET_EXP(pet), pet->short_desc);
 	    send_to_char(buf, ch);
 	}
@@ -3362,7 +3362,7 @@ int pet_shops(struct char_data *ch, int cmd, char *arg)
 	    return eSUCCESS;
 	}
 
-	if (GET_GOLD(ch) < (GET_EXP(pet)*3)) {
+	if (GET_GOLD(ch) < (uint64)(GET_EXP(pet)*3)) {
 	    send_to_char("You don't have enough gold!\n\r", ch);
 	    return eSUCCESS;
 	}
@@ -3508,7 +3508,7 @@ int charon(struct char_data *ch, struct obj_data *obj, int cmd, char *arg,
    }
    
    /* ch typed pay */
-   if(GET_GOLD(ch) < x) {
+   if(GET_GOLD(ch) < (uint64)x) {
       send_to_char("Charon ignores those that cannot afford his services.", ch);
       return eSUCCESS;
    }
@@ -5144,7 +5144,7 @@ int generic_area_guard(struct char_data *ch, struct obj_data *obj, int cmd, char
             break;
       }
 
-      sprintf(buf, "%ld", vict->mobdata->nr);
+      sprintf(buf, "%d", vict->mobdata->nr);
       return do_join(ch, buf, 9);
    }
